481 Readjusted structure turnstile route tests
